Remove compiler warnings on Windows

+++ src/MQTTAsync.c
@@ -2365,7 +2365,7 @@ int MQTTAsync_connect(MQTTAsync handle, const MQTTAsync_connectOptions* options)
 			}
 			else
 			{
-				m->c->will->payloadlen = strlen(options->will->message);
+				m->c->will->payloadlen = (int)strlen(options->will->message);
 				source = (void*)options->will->message;
 			}
 			m->c->will->payload = malloc(m->c->will->payloadlen);
@@ -2446,7 +2446,7 @@ int MQTTAsync_connect(MQTTAsync handle, const MQTTAsync_connectOptions* options)
 	if (options->password)
 	{
 		m->c->password = MQTTStrdup(options->password);
-		m->c->passwordlen = strlen(options->password);
+		m->c->passwordlen = (int)strlen(options->password);
 	}
 	else if (options->struct_version >= 5 && options->binarypwd.data)
 	{
